% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test}
\alias{test}
\title{this a first test whether it is possible to call SimpleITK via Rcpp}
\usage{
test(image, origin = NULL)
}
\arguments{
\item{image}{image of class}

\item{origin}{vector of length image$GetDimension()}
}
\value{
returns the image dimension
}
\description{
this a first test whether it is possible to call SimpleITK via Rcpp. It simply shows how to change the origin of a sitk::Image using SimpleITK's C++ interface.
}
\examples{
require(SimpleITK)
img <- Image(10,10,10,"sitkInt8")
img <-test(img,1:3)
print(img$GetOrigin())
}

